last

+++ app.py
@@ -3,13 +3,15 @@ from flask import Flask, request, abort
 from pathlib import Path
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from sqlalchemy.exc import IntegrityError
 
 BASE_DIR = Path(__file__).parent
 
 app = Flask(__name__)
 app.config['JSON_AS_ASCII'] = False
-app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL').replace("://", "ql://", 1) if os.environ.get('DATABASE_URL') else None \
-                                        or f"sqlite:///{BASE_DIR / 'main.db'}"
+app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL').replace("://", "ql://", 1) if os.environ.get(
+    'DATABASE_URL') else None \
+                         or f"sqlite:///{BASE_DIR / 'main.db'}"
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 migrate = Migrate(app, db, render_as_batch=True)
@@ -59,7 +61,7 @@ def not_found(e):
 def get_object_or_404(model, object_id):
     _object = model.query.get(object_id)
     if _object is None:
-        abort(404, description=f"Object with id={object_id} not found")
+        abort(404, description=f"Author with id={object_id} not found")
 
     return _object
 
@@ -83,7 +85,11 @@ def create_author():
     author_data = request.json
     author = AuthorModel(author_data["name"])
     db.session.add(author)
-    db.session.commit()
+    try:
+        db.session.commit()
+    except IntegrityError:
+        db.session.rollback()
+        return "Author name must be unique", 400
     return author.to_dict(), 201
 
 
@@ -101,7 +107,9 @@ def edit_author(author_id):
 
 @app.route("/authors/<int:author_id>", methods=["DELETE"])
 def delete_author(author_id):
-    author = get_object_or_404(AuthorModel, author_id)
+    author = AuthorModel.query.get(author_id)
+    if author is None:
+        return f"Author with id {author_id} not found.", 404
     db.session.delete(author)
     db.session.commit()
     return author.to_dict(), 201
@@ -121,10 +129,8 @@ def get_quotes():
 
 @app.route("/quotes/<int:id>")
 def get_quote_by_id(id):
-    quote = QuoteModel.query.get(id)
-    if quote:
-        return quote.to_dict()
-    return f"Quote with id {id} not found.", 404
+    quote = get_object_or_404(QuoteModel, id)
+    return quote.to_dict()
 
 
 @app.route("/authors/<int:author_id>/quotes", methods=["POST"])
